+++ RELEASENOTES.md
@@ -2,6 +2,7 @@
 
 - dataKeyPieceOverride functionality is removed
 - paths like 'a.b.c' where 'b' is array is not supported. use a.b[*].c with jsonave.
+- 'dataKeyFnOptions' replaced with 'context' option.
 
 # v1.5.0 - June 12, 2015